Moved split_mps later
